Simplify multi-string encoding
The old implementation works by reversing the string twice, but there
is a far simpler solution:

Strings are 0 terminated: xxx\0yyy\0zzz\0. This makes a simple call to
binary:split/3 with a trim option wrong as it would trim empty
strings. E.g., the input xxx\0\0 would come out as [<<"xxx">>] and not
[<<"xxx">>, <<>>] as desired.

Observe that chopping off the last \0 makes the terminator into a
separator. This yields an optimization because we can create a
sub-binary for everything else and split on that.