Do not track columnar information in the inner loop
epgsql_wire:decode_data/4 is used in an inner loop when doing row
decodes. We track a list here, which is not used for anything but as a
counter. Also, the decoder stack encodes the same value.

Remove the tracking of columns from the inner loop. Normally this
doesn't matter, but this is on the critical decoder path.