Handle errors in the batch command.
If an error occurs while executing a batch
of statements, the error is returned by PG and
all the remaining statements are not executed.
We need to handle that and return appropriate
error message in order to have result list to
match with the statements in the batch.

Before this change, the epgsql user received
an error on unexpected message 90 ("Z") from
the postgres server.
""