fix: reduce log spamming when connection goes down
When a socket goes down due to connectivity issues (e.g.: the connection is
flaky and is broken), then a big crash report is logged.  If this process is supervised
and restarted, that may lead to lots of logs being generated.

Here, `econnrefused` stop reasons are treated as a normal shutdown, so a crash report from
`gen_server` is avoided.